requirements.txt
python-dotenv>=1.0

pydantic==2.9.2
pydantic-settings==2.6.1

boto3
types-boto3 
botocore
boto3-stubs[s3]

pymupdf

langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1

pandas

fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
# pydantic==2.10.3
# pydantic-settings==2.6.1